+++ tests/test_visitors/test_ast/test_classes/test_methods/test_yield_magic_method.py
@@ -1,0 +1,116 @@
+# -*- coding: utf-8 -*-
+
+import pytest
+
+from wemake_python_styleguide.violations.oop import YieldMagicMethodViolation
+from wemake_python_styleguide.visitors.ast.classes import WrongMethodVisitor
+
+method_template = """
+class Test(object):
+    def {0}(self, *args, **kwargs):
+        {1}
+"""
+
+classmethod_template = """
+class Test(object):
+    @classmethod
+    def {0}(cls, *args, **kwargs):
+        {1}
+"""
+
+
+@pytest.mark.parametrize('code', [
+    method_template,
+    classmethod_template,
+])
+@pytest.mark.parametrize('method', [
+    '__init__',
+    '__new__',
+    '__str__',
+    '__aenter__',
+    '__exit__',
+])
+@pytest.mark.parametrize('statement', [
+    'yield',
+    'yield 1',
+    'yield from some()',
+])
+def test_magic_generator(
+    assert_errors,
+    parse_ast_tree,
+    default_options,
+    code,
+    method,
+    statement,
+):
+    """Testing that magic method with `yield` is prohibited."""
+    tree = parse_ast_tree(code.format(method, statement))
+
+    visitor = WrongMethodVisitor(default_options, tree=tree)
+    visitor.run()
+
+    assert_errors(visitor, [YieldMagicMethodViolation])
+
+
+@pytest.mark.parametrize('code', [
+    method_template,
+    classmethod_template,
+])
+@pytest.mark.parametrize('method', [
+    '__init__',
+    '__new__',
+    '__str__',
+    '__aenter__',
+    '__iter__',
+    '__exit__',
+    '__custom__',
+])
+@pytest.mark.parametrize('statement', [
+    'return 1',
+    'print()',
+])
+def test_magic_statement(
+    assert_errors,
+    parse_ast_tree,
+    default_options,
+    code,
+    method,
+    statement,
+):
+    """Testing that magic method with statement is allowed."""
+    tree = parse_ast_tree(code.format(method, statement))
+
+    visitor = WrongMethodVisitor(default_options, tree=tree)
+    visitor.run()
+
+    assert_errors(visitor, [])
+
+
+@pytest.mark.parametrize('code', [
+    method_template,
+    classmethod_template,
+])
+@pytest.mark.parametrize('method', [
+    '__iter__',
+    '__custom__',
+])
+@pytest.mark.parametrize('statement', [
+    'yield',
+    'yield 1',
+    'yield from some()',
+])
+def test_iter_generator(
+    assert_errors,
+    parse_ast_tree,
+    default_options,
+    code,
+    method,
+    statement,
+):
+    """Testing that magic `iter` method with `yield` is allowed."""
+    tree = parse_ast_tree(code.format(method, statement))
+
+    visitor = WrongMethodVisitor(default_options, tree=tree)
+    visitor.run()
+
+    assert_errors(visitor, [])