+++ wemake_python_styleguide/constants.py
@@ -104,6 +104,118 @@ SPECIAL_ARGUMENT_NAMES_WHITELIST: Final = frozenset((
     'mcs',
 ))
 
+#: List of all magic methods from the python docs.
+ALL_MAGIC_METHODS: Final = frozenset((
+    '__new__',
+    '__init__',
+    '__del__',
+
+    '__repr__',
+    '__str__',
+    '__bytes__',
+    '__format__',
+
+    '__lt__',
+    '__le__',
+    '__eq__',
+    '__ne__',
+    '__gt__',
+    '__ge__',
+
+    '__hash__',
+    '__bool__',
+
+    '__getattr__',
+    '__getattribute__',
+    '__setattr__',
+    '__delattr__',
+    '__dir__',
+
+    '__get__',
+    '__set__',
+    '__delete__',
+    '__set_name__',
+
+    '__init_subclass__',
+    '__instancecheck__',
+    '__subclasscheck__',
+    '__class_getitem__',
+
+    '__call__',
+    '__len__',
+    '__length_hint__',
+    '__getitem__',
+    '__setitem__',
+    '__delitem__',
+    '__missing__',
+    '__iter__',
+    '__reversed__',
+    '__contains__',
+
+    '__add__',
+    '__sub__',
+    '__mul__',
+    '__matmul__',
+    '__truediv__',
+    '__floordiv__',
+    '__mod__',
+    '__divmod__',
+    '__pow__',
+    '__lshift__',
+    '__rshift__',
+    '__and__',
+    '__xor__',
+    '__or__',
+    '__radd__',
+    '__rsub__',
+    '__rmul__',
+    '__rmatmul__',
+    '__rtruediv__',
+    '__rfloordiv__',
+    '__rmod__',
+    '__rdivmod__',
+    '__rpow__',
+    '__rlshift__',
+    '__rrshift__',
+    '__rand__',
+    '__rxor__',
+    '__ror__',
+    '__iadd__',
+    '__isub__',
+    '__imul__',
+    '__imatmul__',
+    '__itruediv__',
+    '__ifloordiv__',
+    '__imod__',
+    '__ipow__',
+    '__ilshift__',
+    '__irshift__',
+    '__iand__',
+    '__ixor__',
+    '__ior__',
+    '__neg__',
+    '__pos__',
+    '__abs__',
+    '__invert__',
+    '__complex__',
+    '__int__',
+    '__float__',
+    '__index__',
+    '__round__',
+    '__trunc__',
+    '__floor__',
+    '__ceil__',
+
+    '__enter__',
+    '__exit__',
+
+    '__await__',
+    '__aiter__',
+    '__anext__',
+    '__aenter__',
+    '__aexit__',
+))
+
 #: List of magic methods that are forbidden to use.
 MAGIC_METHODS_BLACKLIST: Final = frozenset((
     # Since we don't use `del`:
@@ -115,12 +227,21 @@ MAGIC_METHODS_BLACKLIST: Final = frozenset((
     '__delattr__',  # since we don't use `delattr()`
 ))
 
-#: List of magic methods that are allowed to be async.
-ASYNC_MAGIC_METHODS_WHITELIST: Final = frozenset((
+#: List of magic methods that are not allowed to be generators.
+YIELD_MAGIC_METHODS_BLACKLIST: Final = ALL_MAGIC_METHODS.difference({
+    # Allowed to be used with ``yield`` keyowrd:
+    '__iter__',
+})
+
+#: List of magic methods that are not allowed to be async.
+ASYNC_MAGIC_METHODS_BLACKLIST: Final = ALL_MAGIC_METHODS.difference({
+    # In order of appearance on
+    # https://docs.python.org/3/reference/datamodel.html#basic-customization
+    # Allowed magic methods are:
     '__anext__',
     '__aenter__',
     '__aexit__',
-))
+})
 
 #: List of nested classes' names we allow to use.
 NESTED_CLASSES_WHITELIST: Final = frozenset((