+++ wemake_python_styleguide/violations/oop.py
@@ -25,6 +25,7 @@ Summary
    WrongSuperCallViolation
    DirectMagicAttributeAccessViolation
    AsyncMagicMethodViolation
+   YieldMagicMethodViolation
    UselessOverwrittenMethodViolation
 
 Respect your objects
@@ -41,6 +42,7 @@ Respect your objects
 .. autoclass:: WrongSuperCallViolation
 .. autoclass:: DirectMagicAttributeAccessViolation
 .. autoclass:: AsyncMagicMethodViolation
+.. autoclass:: YieldMagicMethodViolation
 .. autoclass:: UselessOverwrittenMethodViolation
 
 """
@@ -423,12 +425,12 @@ class AsyncMagicMethodViolation(ASTViolation):
     """
     Forbids to make some magic methods async.
 
-    We forbid to make ``__init__``, ``__eq__``, ``__lt__``, etc async.
     We allow to make ``__anext__``, ``__aenter__``, ``__aexit__`` async.
+    We also allow custom magic methods to be async.
 
     See
-    :py:data:`~wemake_python_styleguide.constants.ASYNC_MAGIC_METHODS_WHITELIST`
-    for the whole list of whitelisted async magic methods.
+    :py:data:`~wemake_python_styleguide.constants.ASYNC_MAGIC_METHODS_BLACKLIST`
+    for the whole list of blacklisted async magic methods.
 
     Reasoning:
         Defining the magic methods as async which are not supposed
@@ -454,8 +456,53 @@ class AsyncMagicMethodViolation(ASTViolation):
 
     """
 
-    error_template = 'Found forbidden async magic method usage: {0}'
+    error_template = 'Found forbidden `async` magic method usage: {0}'
     code = 610
+
+
+@final
+class YieldMagicMethodViolation(ASTViolation):
+    """
+    Forbids to use ``yield`` inside of several magic methods.
+
+    We allow to make ``__iter__`` a generator.
+    See
+    :py:data:`~wemake_python_styleguide.constants.YIELD_MAGIC_METHODS_BLACKLIST`
+    for the whole list of blacklisted generator magic methods.
+
+    Reasoning:
+        Python's datamodel is strict.
+        You cannot make generators from random magic methods.
+        This rule enforces it.
+
+    Solution:
+        Remove ``yield`` from a magic method
+        or rename it to be a custom method.
+
+    Example::
+
+         # Correct:
+        class Example(object):
+            def __init__(self):
+                ...
+
+        # Wrong:
+        class Example(object):
+            def __init__(self):
+                yield 10
+
+    See also:
+        https://docs.python.org/3/reference/datamodel.html
+
+    .. versionadded:: 0.3.0
+    .. versionchanged:: 0.11.0
+    .. versionchanged:: 0.12.0
+
+    """
+
+    error_template = 'Found forbidden `yield` magic method usage'
+    code = 611
+    previous_codes = {439, 435}
 
 
 @final
@@ -488,4 +535,4 @@ class UselessOverwrittenMethodViolation(ASTViolation):
     """
 
     error_template = 'Found useless overwritten method: {0}'
-    code = 611
+    code = 612